loader: move Loader interface into separate package
this will allow to import the Loader interface without importing the
full loader package, which in some cases may end up generating import
cycles
