+++ pkg/datapath/loader/cell.go
@@ -4,14 +4,8 @@
 package loader
 
 import (
-	"context"
-
-	"github.com/vishvananda/netlink"
-
 	"github.com/cilium/cilium/pkg/datapath/linux/sysctl"
-	"github.com/cilium/cilium/pkg/datapath/loader/metrics"
-	"github.com/cilium/cilium/pkg/datapath/tunnel"
-	datapath "github.com/cilium/cilium/pkg/datapath/types"
+	loaderTypes "github.com/cilium/cilium/pkg/datapath/loader/types"
 	"github.com/cilium/cilium/pkg/hive/cell"
 )
 
@@ -21,24 +15,7 @@ var Cell = cell.Module(
 	cell.Provide(NewLoader),
 )
 
-type Loader interface {
-	CallsMapPath(id uint16) string
-	CompileAndLoad(ctx context.Context, ep datapath.Endpoint, stats *metrics.SpanStat) error
-	CompileOrLoad(ctx context.Context, ep datapath.Endpoint, stats *metrics.SpanStat) error
-	CustomCallsMapPath(id uint16) string
-	DetachXDP(iface netlink.Link, bpffsBase, progName string) error
-	DeviceHasTCProgramLoaded(hostInterface string, checkEgress bool) (bool, error)
-	ELFSubstitutions(ep datapath.Endpoint) (map[string]uint64, map[string]string)
-	EndpointHash(cfg datapath.EndpointConfiguration) (string, error)
-	HostDatapathInitialized() <-chan struct{}
-	Reinitialize(ctx context.Context, o datapath.BaseProgramOwner, tunnelConfig tunnel.Config, deviceMTU int, iptMgr datapath.IptablesManager, p datapath.Proxy) error
-	ReinitializeXDP(ctx context.Context, o datapath.BaseProgramOwner, extraCArgs []string) error
-	ReloadDatapath(ctx context.Context, ep datapath.Endpoint, stats *metrics.SpanStat) (err error)
-	RestoreTemplates(stateDir string) error
-	Unload(ep datapath.Endpoint)
-}
-
 // NewLoader returns a new loader.
-func NewLoader(sc sysctl.Sysctl) Loader {
+func NewLoader(sc sysctl.Sysctl) loaderTypes.Loader {
 	return newLoader(sc)
 }
